APIGOV-22993 - move funcs back under MarketplaceMigration
